feat: 完成基础通知逻辑

+++ internal/core/config.go
@@ -44,8 +44,8 @@ type ConfigTracker struct {
 	Code        string          `json:"code" yaml:"code"`                 // 股市编码
 	Frequency   string          `json:"frequency" yaml:"frequency"`       // 轮循频次，最低5s (可选单位 s:秒 m:分钟 h:小时)
 	Continuous  int             `json:"continuous" yaml:"continuous"`     // 轮询频次内连续上涨/下跌x次时弹出提示
-	PriceDiff   float32         `json:"priceDiff" yaml:"priceDiff"`       // 两次轮询【差价】超过该数值时弹出提示
-	PercentDiff float32         `json:"percentDiff" yaml:"percentDiff"`   // 两次轮询【相差百分比】超过该数值时弹出提示
+	PriceDiff   float64         `json:"priceDiff" yaml:"priceDiff"`       // 两次轮询【差价】超过该数值时弹出提示
+	PercentDiff float64         `json:"percentDiff" yaml:"percentDiff"`   // 两次轮询【相差百分比】超过该数值时弹出提示
 	Updates     []*ConfigUpdate `json:"updates,omitempty" yaml:"updates"` // 更新时间段（不传时取默认值）
 }
